app/services/llm_service.py
from langchain_openai import ChatOpenAI
from ..models import StoryParameters, StoryOutline
import os
from typing import List

class LLMService:
    def __init__(self):
        api_key = os.getenv("OPENAI_API_KEY")
        if not api_key:
            raise ValueError("OPENAI_API_KEY environment variable is not set")
        self.llm = ChatOpenAI(model="gpt-4-0125-preview", temperature=0.7, api_key=api_key)

    def generate_outline(self, params: StoryParameters) -> dict:
        # Implement the actual outline generation logic here
        # This is where you'd use self.llm to generate the outline
        # For now, we'll just return a dummy outline
        return {
            "title": "Dummy Outline",
            "setting": ["A magical forest"],
            "characters": ["Hero", "Sidekick"],
            "themes": ["Friendship", "Adventure"],
            "moral": "Believe in yourself",
            "style": "Whimsical",
            "plot": {
                "introduction": ["Once upon a time..."],
                "rising_action": ["The heroes face a challenge..."],
                "climax": ["The final battle..."],
                "resolution": ["They lived happily ever after..."]
            }
        }

    def develop_section(self, section_name: str, section_outline: List[str], additional_context: dict) -> str:
        # Implement the actual section development logic here
        # This is where you'd use self.llm to develop the section
        # For now, we'll just return a dummy developed section
        return f"Developed content for {section_name}: {' '.join(section_outline)}"

llm_service = LLMService()