requirements.txt
fastapi
uvicorn
pydantic
langchain
openai
pinecone-client
sentence-transformers
celery
redis